+++ src/components/DataViews/components/Charts/actions/setChartType.js
