+++ src/components/DataViews/components/Charts/jquery-loader.js
