+++ src/components/DataViews/components/Table/VectorTable/VectorTable.js
