+++ src/components/DataViews/components/Table/VectorTable/helpers/index.js
