+++ src/components/DataViews/components/Table/VectorTable/helpers/prepareCols.js
