+++ src/components/DataViews/components/Table/VectorTable/helpers/prepareVectorRows.js
