+++ src/components/DataViews/components/Table/VectorTable/helpers/scrollbarWidth.js
