+++ src/components/DataViews/components/Table/VectorTable/helpers/setColumnsData.js
