+++ src/components/DataViews/components/Table/VectorTable/helpers/setColumnsValue.js
