+++ src/components/DataViews/components/Table/VectorTable/index.js
