+++ src/components/DataViews/components/ViewLabel.js
@@ -1,0 +1,5 @@
+export function ViewLabel(props) {
+    const { name } = props;
+
+    return <>{name}</>;
+}
