+++ src/components/LabelBrowser/components/Buttons/ShowQuerySettingsButton.js
