+++ src/components/Panel/Panel.js
@@ -5,7 +5,7 @@ import styled from "@emotion/styled";
 import { useLocation } from "react-router-dom";
 import { setRightPanel } from "../../actions/setRightPanel";
 import { setLeftPanel } from "../../actions/setLeftPanel";
-import DataViews from "../DataView/DataViews";
+import DataViews from "../DataViews";
 
 const PanelCont = styled.div`
     display: flex;
@@ -13,60 +13,42 @@ const PanelCont = styled.div`
     flex: 1;
     width: ${(props) => (props.isSplit ? "50%" : "100%")};
 `;
-
+// Panel should have injected data
 export default function Panel(props) {
     const dispatch = useDispatch();
     const { name } = props;
-    const left = useSelector((store) => store.left);
-    const right = useSelector((store) => store.right);
 
+    const panelDispatch = (name, data) => {
+        if (name === "left") return setLeftPanel(data);
+        return setRightPanel(data);
+    };
+
+    const panel = useSelector((store) => store[name]);
     const isSplit = useSelector((store) => store.isSplit);
     const { hash } = useLocation();
 
     useEffect(() => {
         const params = new URLSearchParams(hash.replace("#", ""));
-
-
-        const urlLeft = params.get("left");
-        const urlRight = params.get("right");
-        let leftQueries = [];
-        let rightQueries = [];
-
-        if (urlLeft) {
-            leftQueries = JSON.parse(decodeURIComponent(urlLeft));
+        const panelUrlData = params.get(name);
+        let panelQueries = [];
+        if (panelUrlData) {
+            panelQueries = JSON.parse(decodeURIComponent(panelUrlData));
         }
 
-        if (urlRight) {
-            rightQueries = JSON.parse(decodeURIComponent(urlRight));
-        }
-      
-        if (props.name === "right" && rightQueries.length > 0) {
-            dispatch(setRightPanel(rightQueries));
-        }
-
-        if (props.name === "left" && leftQueries.length > 0) {
-            dispatch(setLeftPanel(leftQueries));
+        if (panelQueries?.length > 0) {
+            dispatch(panelDispatch(name, panelQueries));
         }
     }, []);
 
-    // CHECK ALSO THAT DATAVIEWS IS AN ARRAY 
+    // CHECK ALSO THAT DATAVIEWS IS AN ARRAY
 
-    
-
-
-    const panelData = useMemo(() => {
-        if (name === "left") {
-            return left;
-        } else {
-            return right;
-        }
-    }, [left, right,name]);
+    const panelData = useMemo(() => panel, [panel]);
 
     return (
         <>
             <PanelCont isSplit={isSplit}>
                 <QueriesContainer {...props} queries={panelData} />
-                <DataViews {...props}/>
+                <DataViews {...props} />
             </PanelCont>
         </>
     );
