+++ src/components/QueryItem/QueryItem.js
@@ -62,7 +62,10 @@ const CloseQuery = styled(KeyboardArrowRightOutlinedIcon)`
 function QueryId(props) {
     const [isEditing, setIsEditing] = useState(false);
     const [idText, setIdText] = useState(props.data.idRef);
-
+    const storeTheme = useSelector(({theme})=>theme)
+    const theme = useMemo(()=>{
+       return themes[storeTheme]
+    },[storeTheme])
     function onIdTextChange(e) {
         e.preventDefault();
         const txt = e.target.value;
@@ -89,6 +92,8 @@ function QueryId(props) {
                         background: "transparent",
                         border: "none",
                         outline: "none",
+                        color: theme.textColor
+
                     }}
                     value={idText}
                     placeholder={idText}
